refactor(api): extract Google credential builder into shared module (#103)
Move _build_service and _get_credentials from calendar_tools.py into
app/auth/google_credentials.py as public build_calendar_service() and
get_google_credentials(). Eliminates fragile cross-package coupling
where sync.py imported private functions with pyright suppressions.

Closes #103
